fix(clickhousech@000): Fix for disk cache

See merge request dp/ClickHouse!7783

+++ src/Storages/DiskCache/DiskCacheSettings.cpp
@@ -7,7 +7,7 @@ namespace DB
 void DiskCacheSettings::loadFromConfig(const Poco::Util::AbstractConfiguration & config, const std::string & disk_cache_name)
 {
     std::string config_prefix = fmt::format("{}.{}", prefix, disk_cache_name);
-    lru_max_size = config.getUInt64(config_prefix + ".lru_max_size", 2 * 1024 * 1024 * 1024 * 1024);
+    lru_max_size = config.getUInt64(config_prefix + ".lru_max_size", static_cast<uint64_t>(2) * 1024 * 1024 * 1024 * 1024);
     random_drop_threshold = config.getUInt64(config_prefix + ".random_drop_threshold", 50);
     mapping_bucket_size = config.getUInt64(config_prefix + ".mapping_bucket_size", 5000);
     lru_update_interval = config.getUInt64("disk_cache_strategies.simple.lru_update_interval", 24 * 60 * 60);
